// src/app/auth/model/user.model.ts
export class User {
    email: string;
    password: string;
    role: string;
    confirmPassword: string;
    primaryMobile: string;
    companyName: string;
    companyAddress: string;
    registrationtype: string;
    countryCode: string;
    name: string;
    _id: string;
    MCNumber: string;
    checkboxRemember: string;
}